chore: move files regarding project package structure
