+++ src/main/kotlin/com/mansaeng/diridibackend/controller/UserController.kt
@@ -1,9 +1,9 @@
-package com.mansaeng.diridibackend.user.controller
+package com.mansaeng.diridibackend.controller
 
 import com.mansaeng.diridibackend.config.security.JwtUtil
-import com.mansaeng.diridibackend.user.dto.AuthRequest
-import com.mansaeng.diridibackend.user.dto.AuthResponse
-import com.mansaeng.diridibackend.user.service.UserService
+import com.mansaeng.diridibackend.dto.AuthRequest
+import com.mansaeng.diridibackend.dto.AuthResponse
+import com.mansaeng.diridibackend.service.UserService
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.security.crypto.password.PasswordEncoder
