+++ src/main/kotlin/com/mansaeng/diridibackend/dto/AuthRequest.kt
@@ -1,4 +1,4 @@
-package com.mansaeng.diridibackend.user.dto
+package com.mansaeng.diridibackend.dto
 
 data class AuthRequest(val username: String, val password: String) {
