+++ src/main/kotlin/com/mansaeng/diridibackend/repository/UserRepository.kt
@@ -1,4 +1,4 @@
-package com.mansaeng.diridibackend.user.repository
+package com.mansaeng.diridibackend.repository
 
 import com.mansaeng.diridibackend.entity.user.User
 import org.springframework.data.mongodb.repository.ReactiveMongoRepository
